Refactor the environment structure of the air hockey
